Optimize update_changes

+++ db/20_changes_update.js
@@ -79,8 +79,8 @@ function macroChangesCsv (mode, project, oplProject, csvFeatures, csvUsers, csvM
         echo "   => [\$((\$(date -d now +%s) - \$process_start_t0))s] Accumulate changes table in database"
         ${PSQL} -c "DELETE FROM ${features_table} WHERE ts BETWEEN '${start_ts}' AND '${end_ts}'"
 
-        ${PSQL} -c "CREATE TABLE IF NOT EXISTS ${features_table}_tmp (LIKE ${features_table})"
-        ${PSQL} -c "TRUNCATE TABLE ${features_table}_tmp"
+        ${PSQL} -c "DROP TABLE IF EXISTS ${features_table}_tmp"
+        ${PSQL} -c "CREATE TABLE ${features_table}_tmp (LIKE ${features_table})"
 
         echo "  [\$((\$(date -d now +%s) - \$process_start_t0))s] Copy features"
         ${PSQL} -c "\\COPY ${features_table}_tmp (osmid, version, changeset, action, contrib, ts, userid, tags, geom, tagsfilter) FROM '${csvFeatures}' CSV"
@@ -104,8 +104,8 @@ function macroChangesCsv (mode, project, oplProject, csvFeatures, csvUsers, csvM
         if (csvMembers != null){
             script += `
             echo "  [\$((\$(date -d now +%s) - \$process_start_t0))s] Copy members"
-            ${PSQL} -c "CREATE TABLE IF NOT EXISTS ${members_table}_tmp (LIKE ${members_table})"
-            ${PSQL} -c "TRUNCATE TABLE ${members_table}_tmp"
+            ${PSQL} -c "DROP TABLE IF EXISTS ${members_table}_tmp"
+            ${PSQL} -c "CREATE TABLE ${members_table}_tmp (LIKE ${members_table})"
             ${PSQL} -c "\\COPY ${members_table}_tmp (memberid, osmid, version, pos, role) FROM '${csvMembers}' CSV"
 
             ${PSQL} -v members_table="${members_table}" -v members_table_tmp="${members_table}_tmp" -f "${__dirname}/25_changes_members.sql"
@@ -115,12 +115,17 @@ function macroChangesCsv (mode, project, oplProject, csvFeatures, csvUsers, csvM
     }
 
     if (csvMembers != null){
+        // Features members are only available in features_table and can't only rely on features_table_tmp
+        let features_table_geom = features_table;
+        if (mode == "update"){
+            features_table_geom = `${features_table}_tmp`
+        }
         script += `
         echo "  [\$((\$(date -d now +%s) - \$process_start_t0))s] Building geometries of ways"
-        ${PSQL} -v features_table="${features_table}" -v members_table="${members_table}" -v start_date="'${start_ts}'" -f "${__dirname}/26_changes_geom_ways.sql"
+        ${PSQL} -v features_table="${features_table_geom}" -v features_perm_table="${features_table}" -v members_table="${members_table}" -v start_date="'${start_ts}'" -f "${__dirname}/26_changes_geom_ways.sql"
 
         echo "  [\$((\$(date -d now +%s) - \$process_start_t0))s] Building geometries of relations"
-        ${PSQL} -v features_table="${features_table}" -v members_table="${members_table}" -v start_date="'${start_ts}'" -f "${__dirname}/26_changes_geom_rels.sql"
+        ${PSQL} -v features_table="${features_table_geom}" -v features_perm_table="${features_table}" -v members_table="${members_table}" -v start_date="'${start_ts}'" -f "${__dirname}/26_changes_geom_rels.sql"
         `;
     }
 
@@ -129,15 +134,26 @@ function macroChangesCsv (mode, project, oplProject, csvFeatures, csvUsers, csvM
     ${PSQL} -c "REFRESH MATERIALIZED VIEW ${changes_table}"
 
     echo "  [\$((\$(date -d now +%s) - \$process_start_t0))s] Process usernames"
-    ${PSQL} -c "CREATE TABLE IF NOT EXISTS ${features_table}_users (LIKE pdm_user_names)"
-    ${PSQL} -c "TRUNCATE TABLE ${features_table}_users"
+    ${PSQL} -c "DROP TABLE IF EXISTS ${features_table}_users"
+    ${PSQL} -c "CREATE TABLE ${features_table}_users (LIKE pdm_user_names)"
     ${PSQL} -c "\\COPY ${features_table}_users (username, userid) FROM '${csvUsers}' CSV"
     ${PSQL} -c "INSERT INTO pdm_user_names SELECT * FROM ${features_table}_users ON CONFLICT DO NOTHING"
     ${PSQL} -c "DROP TABLE ${features_table}_users"
 
     if ${HAS_BOUNDARY}; then
         echo "  [\$((\$(date -d now +%s) - \$process_start_t0))s] Populate boundaries"
-        ${PSQL} -v features_table="${features_table}" -v boundary_table="${boundary_table}" -f "${__dirname}/24_changes_boundary.sql"
+        `;
+        // It needs updated geometry in features_table to run and can't only rely on features_table_tmp
+        if (mode == "update"){
+            script += `
+            ${PSQL} -v features_table="${features_table}" -v features_table_tmp="${features_table}_tmp" -v boundary_table="${boundary_table}" -f "${__dirname}/24_changes_boundary_tmp.sql"
+            `;
+        }else{
+            script += `
+            ${PSQL} -v features_table="${features_table}" -v boundary_table="${boundary_table}" -f "${__dirname}/24_changes_boundary.sql"
+            `;
+        }
+        script += `
     fi
 
     if [ -f "${__dirname}/../projects/${project.name}/contribs.sql" ]; then
@@ -213,6 +229,7 @@ var script = `#!/bin/bash
 
 set -e
 mode="$1"
+keep="$2"
 if [[ -z "$mode" ]]; then
     mode="update"
 fi
@@ -391,8 +408,10 @@ Object.values(projects).forEach(project => {
 });
 
 script += `
-    echo "== Removing temp files"
-    rm -f ${CONFIG.WORK_DIR}/*.osh.pbf
+    if [[ "\$keep" = "keep" ]]; then
+        echo "== Removing temp files"
+        rm -f ${CONFIG.WORK_DIR}/*.osh.pbf
+    fi
 ${separator}
 fi
 
@@ -568,8 +587,10 @@ Object.values(projects).forEach(project => {
 });
 
 script += `
-    echo "== Removing temp files"
-    rm -f ${CONFIG.WORK_DIR}/*.osc.*
+    if [[ "\$keep" = "keep" ]]; then
+        echo "== Removing temp files"
+        rm -f ${CONFIG.WORK_DIR}/*.osc.*
+    fi
 fi
 `;
